crypto: msm: crypto driver performance improvement
To minimize spinlock, the qce50 client is assumed that it can only issue
request to qce50 driver one at a time. After a request is issued to
qce50 from qcrypto. Without waiting for completion, other requests can
still be issued until the maximum limit of outstanding requests
in qce50 reaches.

To cut down the chance of udp socket receive buffer overflow the following
schemes are provided -
The number of bunched requests in qce50 is based on the data length of
the current request to cut down delay for smaller packets. In turn,
with smaller delay, the number of completed requests to process in
seq_response() completion function is less.

The scheduling of qcrypto requests are changed from least use to
round robin. This way, the distribution of requests to each engine
is more even. As the result, reordering of completed requests will
be less. Completed requests to handle in completion callback is less
at a time.

Change-Id: I723bac2f9427cddb5248101c9ac3f2b595ad0379
Acked-by: Che-Min Hsieh <cheminh@qti.qualcomm.com>
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>